+++ CHANGELOG.md
@@ -9,6 +9,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 - better error for invalid query selector strings
 
+### Fixed
+
+- a bug where parent interactors were returned within nested
+  methods when using deeper nested methods
+
 ## [0.7.0] - 2018-07-06
 
 ### Changed